+++ src/components/Callout/Callout.types.ts
@@ -1,7 +1,7 @@
-import type { ReactNode } from "react";
+import type { HTMLAttributes, ReactNode } from "react";
 import type { Color } from "../../utils/types";
 
-export interface CalloutProps {
+export interface CalloutProps extends HTMLAttributes<HTMLDivElement> {
   children: ReactNode;
   color?: Color;
   icon?: ReactNode;